requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.8
django-allauth==0.42.0
djangorestframework==3.11.0
filelock==3.0.12
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.0
oauthlib==3.1.0
pep8==1.7.1
Pillow==7.2.0
pipenv==2020.6.2
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
#PyMySQL==0.9.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.17.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
virtualenv==20.0.30
virtualenv-clone==0.5.4
wrapt==1.12.1
zipp==3.1.0
